+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
-## v0.1.0
+## v0.1.2
+
+Fix crash on unknown asset
+Added Edit asset details
+Added Delete asset
+Added Copy receive address button
+
+## v0.1.1
 
 * Initial Release